python/kafka_updater_service/requirements.txt
async-timeout==4.0.2
kafka-python==2.0.2
redis==4.5.5
